# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI营销战略执行智能体",
    "description": "用AI帮小企业把营销战略翻译成每天的执行动作",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/business": {
      "post": {
        "tags": [
          "企业信息"
        ],
        "summary": "Create Business",
        "description": "创建企业信息档案。",
        "operationId": "create_business_api_business_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/business/create": {
      "post": {
        "tags": [
          "企业信息"
        ],
        "summary": "Create Business V3",
        "description": "[V3.0 别名] 创建企业信息档案。",
        "operationId": "create_business_v3_api_business_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/business/info": {
      "get": {
        "tags": [
          "企业信息"
        ],
        "summary": "Get Business Info",
        "description": "[V1.0] 获取企业信息（无路径参数版）。",
        "operationId": "get_business_info_api_business_info_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/business/{business_id}": {
      "get": {
        "tags": [
          "企业信息"
        ],
        "summary": "Get Business",
        "description": "查询企业信息。",
        "operationId": "get_business_api_business__business_id__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessCreateRequest": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name",
            "description": "企业名称"
          },
          "industry": {
            "type": "string",
            "title": "Industry",
            "description": "行业"
          },
          "city": {
            "type": "string",
            "title": "City",
            "description": "所在城市"
          },
          "product_desc": {
            "type": "string",
            "title": "Product Desc",
            "description": "产品/服务描述",
            "default": ""
          },
          "price_range": {
            "type": "string",
            "title": "Price Range",
            "description": "价格区间",
            "default": ""
          },
          "target_customers": {
            "type": "string",
            "title": "Target Customers",
            "description": "目标客户",
            "default": ""
          },
          "competitors": {
            "type": "string",
            "title": "Competitors",
            "description": "竞争对手",
            "default": ""
          },
          "current_channels": {
            "type": "string",
            "title": "Current Channels",
            "description": "当前获客方式",
            "default": ""
          },
          "monthly_revenue": {
            "type": "string",
            "title": "Monthly Revenue",
            "description": "月营业额",
            "default": ""
          },
          "team_size": {
            "type": "string",
            "title": "Team Size",
            "description": "团队规模",
            "default": ""
          },
          "biggest_pain": {
            "type": "string",
            "title": "Biggest Pain",
            "description": "最大痛点",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "industry",
          "city"
        ],
        "title": "BusinessCreateRequest",
        "description": "创建企业信息请求。\n\nParameters\n----------\nbusiness_name : str\n    企业名称（必填）。\nindustry : str\n    所属行业（必填）。\ncity : str\n    所在城市（必填）。\nproduct_desc : str\n    产品/服务描述。\nprice_range : str\n    价格区间。\ntarget_customers : str\n    目标客户群体。\ncompetitors : str\n    竞争对手分析。\ncurrent_channels : str\n    当前获客方式。\nmonthly_revenue : str\n    月营业额。\nteam_size : str\n    团队规模。\nbiggest_pain : str\n    最大营销痛点。"
      },
      "BusinessResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "响应状态码，0表示成功",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "响应数据载荷"
          }
        },
        "type": "object",
        "title": "BusinessResponse",
        "description": "企业信息统一响应格式。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}